\name{POLMM-package}
\alias{POLMM-package}
\alias{POLMM}
\docType{package}
\title{
\packageTitle{POLMM}
}
\description{
\packageDescription{POLMM}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{POLMM}
\packageIndices{POLMM}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{POLMM}

Maintainer: \packageMaintainer{POLMM}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
