% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POLMM_Region.R
\name{POLMM.Region}
\alias{POLMM.Region}
\title{Test for association between genetic variants and an ordinal categorical variable via Proportional Odds Logistic Mixed Model (POLMM-Gene)}
\usage{
POLMM.Region(
  objNull,
  AnnoFile,
  GenoFile,
  GenoFileIndex = NULL,
  OutputFile,
  SparseGRM,
  chrom,
  POLMM.control = NULL
)
}
\arguments{
\item{objNull}{the output of the POLMM_Null_Model() function}

\item{SparseGRM}{an object of class "SparseGRM", check help(getSparseGRM) for more details.}

\item{chrom}{a character to specify chromosome of the markers in GMat. Must be specified unless LOCO = F when fitting the null model.}

\item{PlinkFile}{character, represents the prefix of PLINK input file.}

\item{obj.SNPSet}{an R list with names of variant set and elements of variants IDs in the set}

\item{SKAT.control}{a list of parameters for controlling the POLMM.Gene(). Check 'Details' Section for more details.}
}
\value{
an R matrix with the following elements
\item{ID}{Marker IDs from colnames(GMat)}
\item{chr}{Chromosome name from chrVec}
\item{MAF}{MAFs of the markers}
\item{missing.rate}{Missing rates of the markers}
\item{Stat}{Score statistics}
\item{VarW}{Estimated variance (VarW) from non-mixed model}
\item{VarP}{Estimated variance after adjusting for variance ratio r (VarP = VarW * r)}
\item{beta}{Estimated effect size: Stat / VarP}
\item{pval.norm}{p values calculated from normal approximation}
\item{pval.spa}{p values calculated from saddlepoint approximation}
\item{switch.allele}{a logical value indicating if the REF/ALT alleles were switched, if AF > 0.5, we use GVec = 2-GVec, and then give switch.allele=T. This is useful to estimate the effect direction.}
}
\description{
Test for association between genetic variants and an ordinal categorical variable via Proportional Odds Logistic Mixed Model (POLMM-Gene)
}
\details{
More information about the list of 'SKAT.control'
\itemize{
\item{memoryChunk: Size (Gb) for each memory chunk when reading in Plink file [default=2].}
\item{seed: An integer as a random seed [default=12345678].}
\item{tracenrun: Number of runs for trace estimator [default=30].}
\item{maxiter: Maximum number of iterations used to fit the null POLMM [default=100].}
\item{tolBeta: Positive tolerance: the iterations converge when |beta - beta_old| / (|beta| + |beta_old| + tolBeta) < tolBeta [default=0.001].}
\item{tolTau: Positive tolerance: the iterations converge when |tau - tau_old| / (|tau| + |tau_old| + tolTau) < tolTau [default=0.002].}
\item{tau: Initial value of the variance component (tau) [default=0.2].}
\item{maxiterPCG: Maximum number of iterations for PCG to converge [default=100].}
\item{tolPCG: Positive tolerance for PCG to converge [default=1e-6].}
\item{maxiterEps: Maximum number of iterations for cutpoints estimation [default=100].}
\item{tolEps: Positive tolerance for cutpoints estimation to converge [default=1e-10].}
\item{minMafVarRatio: Minimal value of MAF cutoff to select markers (from Plink file) to estimate variance ratio [default=0.1].}
\item{maxMissingVarRatio: Maximal value of missing rate cutoff to select markers (from Plink file) to estimate variance ratio [default=0.1].}
\item{nSNPsVarRatio: Initial number of the selected markers to estimate variance ratio [default=20], the number will be automatically added by 10 until the coefficient of variantion (CV) of the variance ratio estimate is below CVcutoff.}
\item{CVcutoff: Minimal cutoff of coefficient of variantion (CV) for variance ratio estimation [default=0.0025].}
\item{LOCO: Whether to apply the leave-one-chromosome-out (LOCO) approach [default=TRUE].}
\item{numThreads: Number of threads (CPUs) to use. Only valid if dense GRM is used, default is "auto", that is, RcppParallel::defaultNumThreads() [default="auto"].}
\item{stackSize: Stack size (in bytes) to use for worker threads. For more details, check help(RcppParallel::setThreadOptions) [default="auto"].}
\item{grainSize: Grain size of a parallel algorithm sets a minimum chunk size for parallelization. In other words, at what point to stop processing input on separate threads [default=1].}
\item{minMafGRM: Minimal value of MAF cutoff to select markers (from Plink files) to construct dense GRM [default=0.01].}
\item{maxMissingGRM: Maximal value of missing rate to select markers (from Plink files) to construct dense GRM [default=0.1].}
\item{showInfo: Whether to show more detailed information for trouble shooting [default=TRUE].}
\item{onlyCheckTime: Not fit the null model, only check the computation time of reading Plink files and running 30 KinbVec() functions [default=FALSE].}
}
}
\examples{
famFile = system.file("extdata", "nSNPs-10000-nsubj-1000-ext.fam", package = "POLMM")
GenoFile = gsub("-ext.fam", "-ext.bed", famFile)
AnnoFile = system.file("extdata", "AnnoFile.txt", package = "POLMM")
OutputFile = gsub("AnnoFile","OutputFile",AnnoFile)
SparseGRMFile = system.file("SparseGRM", "SparseGRM.RData", package = "POLMM")
load(SparseGRMFile)
objNullFile = system.file("objNull.RData", package = "POLMM")
load(objNullFile)
chrom = 1

OUTPUT = POLMM.Region(objNull, AnnoFile, GenoFile, GenoFileIndex = NULL, OutputFile, 
                      SparseGRM, chrom, POLMM.control = list(max_maf_region = 0.5))
     
}
