% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare_Null_Model.R
\name{getSparseGRMParallel}
\alias{getSparseGRMParallel}
\title{Please run this function before running getSparseGRM().}
\usage{
getSparseGRMParallel(
  chrParallel,
  partParallel,
  nPartsGRM,
  PlinkFile,
  tempDir = NULL,
  subjData = NULL,
  minMafGRM = 0.01,
  maxMissingGRM = 0.1,
  threadNum = 8
)
}
\arguments{
\item{chrParallel}{a character of chromosome number (e.g. 1, 2, ..., 22), better autosomes.}

\item{partParallel}{a numeric value (from 1 to nPartsGRM)}

\item{nPartsGRM}{a numeric value (e.g. 250): GCTA software can split subjects to multiple parts. For UK-Biobank analysis, it is recommanded to use 250 parts.}

\item{PlinkFile}{a path to Plink files. The current version (gcta_1.93.1beta) of gcta software does not support difference prefix names for bim, bed and fam files.}

\item{tempDir}{a path to store temp files from getSparseGRMParallel(). This should be consistent to the input of getSparseGRM(). Default is system.file("SparseGRM", "temp", package = "POLMM").}

\item{subjData}{a character vector to specify subject IDs to select (i.e. IID). Default is NULL, i.e. all subjects are selected in sparse GRM. If the number of subjects is less than 1,000, the GRM estimation might not be accurate.}

\item{minMafGRM}{Minimal value of MAF cutoff to select markers (from Plink files) to construct GRM.}

\item{maxMissingGRM}{Maximal value of missing rate to select markers (from Plink files) to construct GRM.}

\item{threadNum}{Number of threads (CPUs) to use.}
}
\description{
Please run this function before running getSparseGRM(). We strongly suggest using parallel computing for different pairs of (chrParallel, partParallel).
}
\examples{
## Please check help(getSparseGRM) for example codes.
}
