% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{POLMM}
\alias{POLMM}
\title{Test for association between genotype and an ordinal categorical variable via Proportional Odds Logistic Mixed Model (POLMM)}
\usage{
POLMM(GMat, obj_Null, chrVec, minMAF)
}
\arguments{
\item{GMat}{a numeric genotype matrix with each row as an individual and each column as a marker. Column names is required}

\item{obj_Null}{output object of function FitNull}

\item{chrVec}{a character or vector to specify the markers chromosome}

\item{minMAF}{lower cutoff of genetic variants minor allele frequencies (MAFs) for analysis.}
}
\value{
an R matrix with the following elements
\item{SNPID}{SNP IDs from colnames(GMat)}
\item{MAF}{Minor allele frequencies}
\item{Stat}{Score statistics}
\item{VarW}{Estimated variance (VarW) from non mixed effect model}
\item{VarP}{Estimated variance after adjusting for variance ratio r (VarP = VarW * r)}
\item{pval.norm}{p values calculated from normal approximation}
\item{pval.spa}{p values calculated from saddlepoint approximation}
}
\description{
Test for association between genotype and an ordinal categorical variable via Proportional Odds Logistic Mixed Model (POLMM)
}
\examples{
## The example files use plink files with 10000 markers and 1000 subjects. 
## For real data analysis, we recommend >= 100000 markers.
famFile = system.file("extdata", "nSNPs-10000-nsubj-1000-ext.fam", package = "POLMM")
PlinkFile = gsub("-ext.fam","-ext",famFile)
dataFile = system.file("extdata", "POLMM_data.csv", package = "POLMM")

famData = data.table::fread(famFile)
egData = data.table::fread(dataFile)

## Fit the null POLMM 
objNull = POLMM_Null_Model(as.factor(outcome)~Cova1+Cova2, data=egData, PlinkFile = PlinkFile, subjData = egData$IID, subjPlink = famData$V2)

## If control$seed is not changed, objNull$tau should be 0.705
objNull$tau

## when using SAIGE.POLMM, chrVec should be from
names(objNull$LOCOList)

GMat = matrix(rbinom(10000,2,0.3),1000,10)
colnames(GMat) = paste0("rs",1:10)
outPOLMM = POLMM(GMat, objNull, "1", 0.001)
}
