% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{POLMM_Null_Model}
\alias{POLMM_Null_Model}
\title{Fit a Proportional Odds Logistic Mixed Model (POLMM) for ordinal categorical data analysis}
\usage{
POLMM_Null_Model(formula, data, PlinkFile, subjData, subjPlink,
  subjMatch = FALSE, GMatRatio, control)
}
\arguments{
\item{formula}{an object of class "formula" to be passed to function ordinal::clm().}

\item{data}{an optional data frame containing the variables in the null model. If not specified, the variables are taken from 'environment(formula)'}

\item{PlinkFile}{a path to Plink files for Genetic Relationship Matrix (GRM).}

\item{subjData}{a character vector to specify subject IDs in 'formula' and 'data'.}

\item{subjPlink}{a character vector to specify subject IDs in 'PlinkFile'.}

\item{subjMatch}{a logical value to indicate if subjData is the same as subjPlink. If TRUE, 'subjData' and 'subjPlink' are not needed any more. Default is FALSE.}

\item{GMatRatio}{only required if control$LOCO=F. a numeric matrix (NA for missin data) with genotype of multiple markers (>= 100 markers) to calculate variance ratio (r=VarP/VarW, see 'Details' Section). Each row is for one individual (the same order as subjData) and each column is for one marker. We recommand that these markers should be uncorrelated with the markers in PlinkFile.}

\item{control}{a list of parameters for controlling the fitting process. More details can be seen in 'Details' Section.}
}
\value{
an R object of 'POLMM_NULL' with the following elements, J is number of categories.
\item{N}{Number of individuals in analysis.}
\item{M}{Number of markers in Plink file.}
\item{Cova}{A stacked matrix (N(J-1) x p) of covariates; p is number of covariates.}
\item{beta}{Fixed effect: estimated parameters for covariates.}
\item{tau}{Variance component. NOTE: usually underestimated due to Penalized Quasi-Likelihood (PQL).}
\item{bVec}{Random effect: estimated to characterize sample relatedness}
\item{eps}{Cutpoints for different categories}
\item{eta}{Linear predicators for all individuals}
\item{yVec}{A vector of the categorical phenotypes.}
\item{muMat}{A matrix (N x J), probability in each category.}
\item{YMat}{A matrix (N x (J-1)), working variable in each category.}
\item{subjIDs}{A vector of individual IDs in null model}
\item{LOCOList}{Objects for Step 2 calculation. Might be large if sample size is large and number of chromosomes is large. For example, when analyzing UK Biobank with xx individuals and xx chromosomes, this object is xx Gb.}
\item{controlList}{List of control arguments. More details can be seen in 'Details' Section.}
\item{sessionInfo}{Information of R session when analysis}
\item{time}{Time when null model fitting is finished}
}
\description{
Fit a POLMM with a response varaible of ordinal categorical outcome. To adjust for sample relatedness, we use a random effect with a variance estiamted from Genetic Relationship Matrix (GRM).
}
\details{
Variance VarP should be used to calibrate p values. However, it is computational expensive to calculate VarP for all markers. Hence, we first estimate variance ratio r = VarP / VarW, and then calculate VarW for all markers and estimate VarP = VarW * r. 
\itemize{
\item{VarP: variance estimated from mixed effect model.}
\item{VarW: variance estimated from non-mixed effect model.}
}
More information about the list of 'control'
\itemize{
\item{memoryChunk: Size (Gb) for each memory chunk when reading in Plink file [default=2].}
\item{seed: An integer as random seed [default=12345678].}
\item{tracenrun: Number of runs for trace estimator [default=30].}
\item{maxiter: Maximum number of iterations used to fit the null POLMM [default=100].}
\item{tolBeta: Positive tolerance for fixed effect coefficients; the iterations converge when |beta - beta_old| / (|beta| + |beta_old| + tolBeta) < tolBeta [default=0.001].}
\item{tolTau: Positive tolerance for random effect variance component (tau); the iterations converge when |tau - tau_old| / (|tau| + |tau_old| + tolTau) < tolTau [default=0.002].}
\item{tau: Initial value of variance component (tau) [default=0.1].}
\item{maxiterPCG: Maximum number of iterations for PCG to converge [default=100].}
\item{tolPCG: Positive tolerance for PCG to converge [default=1e-6].}
\item{maxiterEps: Maximum number of iterations for cutpoints estimation [default=100].}
\item{tolEps: Positive tolerance for cutpoints estimation to converge [default=1e-10].}
\item{minMafVarRatio: Minimal value of MAF cutoff to select markers (from Plink file) to estimate variance ratio [default=0.1].}
\item{nSNPsVarRatio: Initial number of selected markers to estimate variance ratio [default=20], the number will be automatically added by 10 until the coefficient of variantion (CV) of the variance ratio estimate is below CVcutoff.}
\item{CVcutoff: Minimal cutoff of coefficient of variantion (CV) for variance ratio estimation [default=0.0025].}
\item{LOCO: Whether to apply the leave-one-chromosome-out (LOCO) approach, if FALSE, 'GMatRatio' is required [default=TRUE].}
\item{numThreads: Number of threads (CPUs) to use, default is "auto", that is, RcppParallel::defaultNumThreads() [default="auto"].}
\item{stackSize: Stack size (in bytes) to use for worker threads. The default used for "auto" is 2MB on 32-bit systems and 4MB on 64-bit systems (note that this parameter has no effect on Windows) [default="auto"].}
\item{grainSize: [default=1]}
\item{minMafGRM: Minimal value of MAF cutoff to select markers (from Plink file) for GRM [default=0.01].}
\item{maxMissingGRM: Maximal value of missing rate to select markers (from Plink file) for GRM [default=0.1].}
\item{hidePCGInfo: [default=TRUE]}
\item{onlyCheckTime: [default=FALSE]}
}
}
\examples{
## We use a Plink file with 10,000 markers and 1,000 subjects to constract GRM for demonstration. 
## For real data analysis, we recommend >= 100,000 common markers (MAF > 0.05 or 0.01).
## Selection of the common markers is similar as in Principle Components Analysis (PCA).
famFile = system.file("extdata", "nSNPs-10000-nsubj-1000-ext.fam", package = "POLMM")
PlinkFile = gsub("-ext.fam","-ext",famFile)
dataFile = system.file("extdata", "POLMM_data.csv", package = "POLMM")

famData = data.table::fread(famFile)
egData = data.table::fread(dataFile)

## Fit the null POLMM 
objNull = POLMM_Null_Model(as.factor(outcome)~Cova1+Cova2, 
                           data=egData, PlinkFile = PlinkFile, subjData = egData$IID, subjPlink = famData$V2)

## If control$seed is not changed, objNull$tau should be 0.7357
objNull$tau

## If you want more accurate parameter estimation, smaller tolBeta/tolTau can be used. 
objNull = POLMM_Null_Model(as.factor(outcome)~Cova1+Cova2, 
                           data=egData, PlinkFile = PlinkFile, subjData = egData$IID, subjPlink = famData$V2,
                           control=list(tolTau=0.001))  # Default toTau = 0.002
objNull$tau  # 0.7368

## This is for me to check the computational efficiency
objNull = POLMM_Null_Model(as.factor(outcome)~Cova1+Cova2, 
                           data=egData, PlinkFile = PlinkFile, subjData = egData$IID, subjPlink = famData$V2,
                           control=list(onlyCheckTime = T, numThreads = 4))  # Default nThreads = 0, that is, RcppParallel::defaultNumThreads()
  
}
