% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POLMM_Null_Model.R
\name{POLMM_Null_Model}
\alias{POLMM_Null_Model}
\title{Fit a Proportional Odds Logistic Mixed Model (POLMM) for ordinal categorical data analysis}
\usage{
POLMM_Null_Model(
  formula,
  data,
  subjData,
  PlinkFile,
  bimFile,
  bedFile,
  famFile,
  SparseGRM,
  GMatRatio,
  control
)
}
\arguments{
\item{formula}{a formula of the form response ~ predictors to be passed to function ordinal::clm(). The response should be an ordered factor and the predictors should not include an intercept term.}

\item{data}{an optional data frame containing the variables in the null model. If not specified, the variables are taken from 'environment(formula)'.}

\item{subjData}{a character vector to specify subject IDs in 'formula' and 'data'.}

\item{PlinkFile}{a path to Plink files, to construct dense GRM or to extract markers to estimate variant ratio.}

\item{bimFile}{a path to Plink bim file. Default is paste0(PlinkFile,".bim") if not specified.}

\item{bedFile}{a path to Plink bed file. Default is paste0(PlinkFile,".bed") if not specified.}

\item{famFile}{a path to Plink fam file. Default is paste0(PlinkFile,".fam") if not specified.}

\item{SparseGRM}{an object of class "SparseGRM", check help(getSparseGRM) for more details.}

\item{GMatRatio}{an object of class "GMatRatio", usually not used.}

\item{control}{a list of parameters for controlling the fitting process. Check 'Details' Section for more details.}
}
\value{
an R object of 'POLMM_NULL' with the following elements, J is number of categories.
\item{N}{Number of subjects in analysis.}
\item{M}{Number of markers in Plink files.}
\item{Cova}{A stacked matrix (N(J-1) x p) of covariates; p is number of covariates.}
\item{beta}{Fixed effect: estimated parameters for covariates.}
\item{tau}{Variance component. NOTE: usually underestimated due to Penalized Quasi-Likelihood (PQL).}
\item{bVec}{Random effect: estimated to characterize sample relatedness}
\item{eps}{Cutpoints for different categories}
\item{eta}{Linear predicators for all subjects}
\item{yVec}{A vector of the categorical phenotypes.}
\item{muMat}{A matrix (N x J), probability in each category.}
\item{YMat}{A matrix (N x (J-1)), working variable in each category.}
\item{subjIDs}{A vector of subject IDs in null model}
\item{LOCOList}{Main objects to be passed to step 2. Might be large if the sample size is large and number of chromosomes is large.}
\item{controlList}{List of control arguments. More details can be seen in 'Details' Section.}
\item{sessionInfo}{Information of R session when analyzing.}
\item{time}{Time when null model fitting is finished.}
}
\description{
Fit a POLMM with a response varaible of ordinal categorical outcome. If the outcome is binary, POLMM degenerates to a logistic mixed model and the package performs similarly as SAIGE. 
Two options are provided to adjust for sample relatedness: 
If argument SparseGRM is specified, we use a sparse Genetic Relationship Matrix (GRM), check help(getSparseGRM) for more details; 
otherwise, we use a dense GRM estimated from the markers in plink files.
}
\details{
More information about the list of 'control'
\itemize{
\item{memoryChunk: Size (Gb) for each memory chunk when reading in Plink file [default=2].}
\item{seed: An integer as a random seed [default=12345678].}
\item{tracenrun: Number of runs for trace estimator [default=30].}
\item{maxiter: Maximum number of iterations used to fit the null POLMM [default=100].}
\item{tolBeta: Positive tolerance: the iterations converge when |beta - beta_old| / (|beta| + |beta_old| + tolBeta) < tolBeta [default=0.001].}
\item{tolTau: Positive tolerance: the iterations converge when |tau - tau_old| / (|tau| + |tau_old| + tolTau) < tolTau [default=0.002].}
\item{tau: Initial value of the variance component (tau) [default=0.2].}
\item{maxiterPCG: Maximum number of iterations for PCG to converge [default=100].}
\item{tolPCG: Positive tolerance for PCG to converge [default=1e-6].}
\item{maxiterEps: Maximum number of iterations for cutpoints estimation [default=100].}
\item{tolEps: Positive tolerance for cutpoints estimation to converge [default=1e-10].}
\item{minMafVarRatio: Minimal value of MAF cutoff to select markers (from Plink file) to estimate variance ratio [default=0.1].}
\item{maxMissingVarRatio: Maximal value of missing rate cutoff to select markers (from Plink file) to estimate variance ratio [default=0.1].}
\item{nSNPsVarRatio: Initial number of the selected markers to estimate variance ratio [default=20], the number will be automatically added by 10 until the coefficient of variantion (CV) of the variance ratio estimate is below CVcutoff.}
\item{CVcutoff: Minimal cutoff of coefficient of variantion (CV) for variance ratio estimation [default=0.0025].}
\item{LOCO: Whether to apply the leave-one-chromosome-out (LOCO) approach [default=TRUE].}
\item{numThreads: Number of threads (CPUs) to use. Only valid if dense GRM is used, default is "auto", that is, RcppParallel::defaultNumThreads() [default="auto"].}
\item{stackSize: Stack size (in bytes) to use for worker threads. For more details, check help(RcppParallel::setThreadOptions) [default="auto"].}
\item{grainSize: Grain size of a parallel algorithm sets a minimum chunk size for parallelization. In other words, at what point to stop processing input on separate threads [default=1].}
\item{minMafGRM: Minimal value of MAF cutoff to select markers (from Plink files) to construct dense GRM [default=0.01].}
\item{maxMissingGRM: Maximal value of missing rate to select markers (from Plink files) to construct dense GRM [default=0.1].}
\item{showInfo: Whether to show more detailed information for trouble shooting [default=TRUE].}
\item{onlyCheckTime: Not fit the null model, only check the computation time of reading Plink files and running 30 KinbVec() functions [default=FALSE].}
}
}
\examples{
## We use a Plink file with 10,000 markers and 1,000 subjects to constract GRM for demonstration. 
## For real data analysis, we recommend >= 100,000 common markers (MAF > 0.05 or 0.01).
## Selection of the common markers is similar as in Principle Components Analysis (PCA).
famFile = system.file("extdata", "nSNPs-10000-nsubj-1000-ext.fam", package = "POLMM")
PlinkFile = gsub("-ext.fam","-ext",famFile)
dataFile = system.file("extdata", "POLMM_data.csv", package = "POLMM")

egData = data.table::fread(dataFile)

## Fit the null POLMM using the Dense GRM
objNull = POLMM_Null_Model(as.factor(outcome)~Cova1+Cova2, 
                           data=egData, PlinkFile = PlinkFile, subjData = egData$IID)

## If control$seed is not changed, objNull$tau should be 0.7353
objNull$tau

## Fit the null POLMM using the Sparse GRM
SparseGRMFile = system.file("extdata", "SparseGRM.RData", package = "POLMM")
load(SparseGRMFile)   ## check getSparseGRM() for more details about how to make an R object of "SparseGRM" using Plink files. 
objNull = POLMM_Null_Model(as.factor(outcome)~Cova1+Cova2, 
                           SparseGRM = SparseGRM,
                           data=egData, PlinkFile = PlinkFile, subjData = egData$IID)
                           
## If control$seed is not changed, objNull$tau should be 0.8506
objNull$tau


## If you want more accurate parameter estimation, smaller tolBeta/tolTau can be used. 
objNull = POLMM_Null_Model(as.factor(outcome)~Cova1+Cova2, 
                           data=egData, PlinkFile = PlinkFile, subjData = egData$IID,
                           control=list(tolTau=0.001))  # Default toTau = 0.002
objNull$tau  # 0.7368

## This is for me to check the computational efficiency
objNull = POLMM_Null_Model(as.factor(outcome)~Cova1+Cova2, 
                           data=egData, PlinkFile = PlinkFile, subjData = egData$IID,
                           control=list(onlyCheckTime = T, numThreads = 4))  # Default nThreads = 0, that is, RcppParallel::defaultNumThreads()
  
}
